backend/services/15_auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12